submittal-builder/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-cloud-aiplatform==1.38.0
PyPDF2==3.0.1
reportlab==4.0.4
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
jinja2==3.1.2
